Tests/LibWeb/Text/expected/Crypto/SubtleCrypto-aescbc.txt

exported key, raw: 32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47
exported key, jwk: {"kty":"oct","key_ops":["encrypt","decrypt"],"alg":"A128CBC","ext":true,"k":"ICEiIyQlJicoKSorLC0uLw"}
ciphertext: 193,138,47,224,83,254,87,179,130,27,206,126,144,28,38,111,123,45,177,85,53,201,206,231,38,234,7,146,184,159,64,117,220,3,201,210,82,171,14,237,82,53,162,239,16,172,50,72
plaintextRoundtrip: 87,101,108,108,72,101,108,108,111,70,114,105,101,110,100,115,16,16,16,16,16,16,255,16,16,16,16,16,16,16,16,16
odd padding results in: OperationError
broken padding results in: OperationError
